+++ test/sandbox/app/components/action_menu_panel_wrapper_component.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class ActionMenuPanelWrapperComponent < ViewComponent::Base
+  def call
+    render "shared/action_menu_panel"
+  end
+end